finance_app/cards/requirements.txt
belvo-python
python-dotenv
requests
